+++ .env
@@ -1,1 +1,1 @@
-RUST_LOG="info,srtm=debug"
+HGT_TILES_FOLDER="/home/aurelien/Documents/data/hgt/"
